+++ modules/data_structures/source/graph.cpp
@@ -48,7 +48,7 @@ std::vector <int> Graph::dfs(int root) {
 }
 
 int Graph::Dijkstra(int from, int to) {
-    const int INF = 1e9;
+    const int INF = static_cast<int>(1e9);
     std::vector <int> d(n, INF), p(n);
     d[from] = 0;
     std::set <std::pair<int, int> > s;
